requirements.txt
fastapi
uvicorn
sqlalchemy
langchain-community
python-dotenv
openai
pandas
